salon_backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
mongoengine==0.27.0